begin clearing errors and warning
80 %  type-based:
 <Qxxx> added,
 Qt:: added,
 Qyyy::ENUM changed,
 QVector overlaods deleted (is alias of QList),
 QzzzLiteral are now simply QzzzString,
 QFileInfo constructor around paths,
 comparison with letters: 'A' (not "A"),
 some new methods ("secsSinceEpoch" instead of "time_t", "setEncoding" instead of "setCodec",
 ...
some explicit casts (5%)
some msvc #pragmas (5%)
logic adjusment File Dialog (3%),
logic correction SQL (7%) (prepared statement placeholders have no API to get them!)

+++ src/fs/db/airwayresolver.cpp
@@ -114,7 +114,7 @@ uint qHash(const AirwayResolver::AirwaySegment& segment)
          static_cast<unsigned int>(segment.toWaypointId) ^
          static_cast<unsigned int>(segment.minAlt) ^
          static_cast<unsigned int>(segment.maxAlt) ^
-         qHash(segment.type);
+         (uint)qHash(segment.type);
 }
 
 AirwayResolver::AirwayResolver(sql::SqlDatabase *sqlDb, atools::fs::ProgressHandler& progress)
